refactor: 공통 테스트 로또 클래스 병합 및 로또 검증 클래스 분리
