+++ src/Lotto.js
@@ -1,32 +1,11 @@
-import CommonValidator from './validator/Common';
+import LottoValidator from './validator/Lotto.js';
 
 class Lotto {
   #numbers;
 
   constructor(numbers) {
-    this.#validate(numbers);
-    this.#duplicate(numbers);
-    this.#validateRange(numbers);
+    LottoValidator.validate(numbers);
     this.#numbers = this.#sorted(numbers);
-  }
-
-  #validate(numbers) {
-    if (numbers.length !== 6) {
-      throw new Error('[ERROR] 로또 번호는 6개여야 합니다.');
-    }
-  }
-
-  // TODO: 추가 기능 구현
-  #duplicate(numbers) {
-    const uniqueNumbers = new Set(numbers);
-
-    if (uniqueNumbers.size !== numbers.length) {
-      throw new Error('[ERROR] 로또 번호는 중복될 수 없습니다.');
-    }
-  }
-
-  #validateRange(numbers) {
-    numbers.forEach(number => CommonValidator.validateLottoNumber(number));
   }
 
   #sorted(numbers) {
